whatthefile#109 Añadido Firefox a la lista de compatibilidad de navegadores para obtener el historial
