+++ tests/modules/browserhistorytest.py
@@ -55,6 +55,26 @@ class BrowserHistoryTest(unittest.TestCase):
         self.assertEqual(len(result["visites"]), 25)
         self.assertEqual(len(result["searchs"]), 0)
 
+    def test_firefox_history_v1(self):
+        path = "./tests/examples/browsers/firefox/places.sqlite"
+        target_file = TargetFile(path)
+        module = Constructor()
+        self.assertTrue(module.is_valid_for(target_file))
+        result = module.run(target_file, {})
+        self.assertEqual(len(result["downloads"]), 0)
+        self.assertEqual(len(result["visites"]), 1)
+        self.assertEqual(len(result["searchs"]), 0)
+
+    def test_firefox_history_v2(self):
+        path = "./tests/examples/browsers/firefox/places_new.sqlite"
+        target_file = TargetFile(path)
+        module = Constructor()
+        self.assertTrue(module.is_valid_for(target_file))
+        result = module.run(target_file, {})
+        self.assertEqual(len(result["downloads"]), 1)
+        self.assertEqual(len(result["visites"]), 34)
+        self.assertEqual(len(result["searchs"]), 0)
+
     def test_invalid_file(self):
         path = "./tests/examples"
         target_file = TargetPath(path)
